backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
redis==5.0.1
celery==5.3.4
elasticsearch==8.11.0
sentence-transformers==3.3.1
numpy==1.24.3
django-cors-headers==4.3.1
gunicorn==21.2.0